+++ forui/lib/src/widgets/toast/toaster_style.dart
@@ -1,3 +1,5 @@
+import 'dart:ui';
+
 import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 
@@ -129,6 +131,14 @@ class FToastStyle with Diagnosticable, _$FToastStyleFunctions {
   @override
   final Curve transitionCurve;
 
+  /// The toast's swipe completion animation duration. Defaults to 150ms.
+  @override
+  final Duration swipeCompletionDuration;
+
+  /// The toast's swipe completion animation curve. Defaults to [Curves.easeInCubic].
+  @override
+  final Curve swipeCompletionCurve;
+
   /// The toast's constraints. Defaults to `BoxConstraints(maxHeight: 250, maxWidth: 400)`.
   @override
   final BoxConstraints constraints;
@@ -136,6 +146,34 @@ class FToastStyle with Diagnosticable, _$FToastStyleFunctions {
   /// The toast's decoration.
   @override
   final BoxDecoration decoration;
+
+  /// An optional background filter. This only takes effect if the [decoration] has a transparent or translucent
+  /// background color.
+  ///
+  /// This is typically combined with a transparent/translucent background to create a glassmorphic effect.
+  ///
+  /// There will be a flicker after the toast's fade-in entrance when a blur background filter is applied. This is due to
+  /// https://github.com/flutter/flutter/issues/31706.
+  ///
+  /// ## Examples
+  /// ```dart
+  /// // Blurred
+  /// ImageFilter.blur(sigmaX: 5, sigmaY: 5);
+  ///
+  /// // Solid color
+  /// ColorFilter.mode(Colors.white, BlendMode.srcOver);
+  ///
+  /// // Tinted
+  /// ColorFilter.mode(Colors.white.withValues(alpha: 0.5), BlendMode.srcOver);
+  ///
+  /// // Blurred & tinted
+  /// ImageFilter.compose(
+  ///   outer: ImageFilter.blur(sigmaX: 5, sigmaY: 5),
+  ///   inner: ColorFilter.mode(Colors.white.withValues(alpha: 0.5), BlendMode.srcOver),
+  /// );
+  /// ```
+  @override
+  final ImageFilter? backgroundFilter;
 
   /// The toast content's padding. Defaults to `EdgeInsets.all(16)`.
   @override
@@ -177,8 +215,11 @@ class FToastStyle with Diagnosticable, _$FToastStyleFunctions {
     this.entranceExitOpacity = 0.0,
     this.transitionDuration = const Duration(milliseconds: 400),
     this.transitionCurve = Curves.easeOutCubic,
+    this.swipeCompletionDuration = const Duration(milliseconds: 150),
+    this.swipeCompletionCurve = Curves.easeInCubic,
     this.constraints = const BoxConstraints(maxHeight: 250, maxWidth: 400),
     this.padding = const EdgeInsets.all(16),
+    this.backgroundFilter,
     this.iconSpacing = 10,
     this.titleSpacing = 1,
     this.suffixSpacing = 12,